feat: rename coq to rocq
