+++ lua/visimp/languages/coq.lua
@@ -1,13 +1,3 @@
-local layer = require 'visimp.layer'
-local L = layer.new_layer 'coq'
-
-function L.packages()
-  return { 'whonore/Coqtail' }
-end
-
-function L.load()
-  vim.cmd 'packadd Coqtail'
-  layer.to_vimscript_config(L, 'coqtail_', true)
-end
-
-return L
+local coqLayer = vim.deepcopy(require 'visimp.languages.rocq')
+coqLayer.identifier = 'coq'
+return coqLayer
